Change default to have corosync identify nic
Currently the principle charm explicitly states what nic should be
used for hosting a vip. But if nic names and/or network assignments
are not consistent across units of an application this will break.
This change removes the default behaviour of supplying the nic and
instead defers it to corosync/pacemaker, this removes the requirement
that nics must be consistent across units. If a nic or subnet is
explicitly supplied via principle charm config then it is still
passed on explicitly.

Change-Id: I9dfe6b01b0d09847bac499e7e30360b81960c93a