Do not spicy iface when creating ha vip resource
Pacemaker VIP resources are able to automatically detect and
configure correct iface and netmask parameters based on local
configuration of the unit. The hacluster interface will use
this approach if nic and netmask are omitted when setting up the
vip. If nic and netmask cannot be automatically detected then
fallback to the old behaviour. If new behaviour is being used
then ensure the old vip resource name is removed.

Change-Id: Id2fbb71328572bada939d432030d9d7ab4814f1b